Extract GTModelGeometry test method to class
